pyproject.toml
[project]
name = "ensemble-cam"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13.2"
dependencies = [
    "jupyterlab>=4.0.0",
    "datasets>=2.14.0",
    "huggingface-hub>=0.16.0",
    "tqdm>=4.60.0",
    "pillow>=10.0.0",
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "scikit-learn>=1.3.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "tensorboard>=2.13.0",
]
